added editing option to org Details and Address

+++ com.example.e4.rcp.todo/src/com/example/e4/rcp/tables/Address.java
@@ -11,7 +11,7 @@ import javax.persistence.Table;
 public class Address implements DatabaseAccess{
 	
 	@Id @GeneratedValue(strategy= GenerationType.AUTO)
-	private int addId;
+	private long addId;
 	private String address;
 	private String state;
 	private Integer postalCode;
